pyproject.toml
[tool.poetry]
authors = ["Buried-In-Code <BuriedInCode@tuta.io>"]
classifiers = [
	"Development Status :: 5 - Alpha",
	"Environment :: Console",
	"Intended Audience :: End Users/Desktop",
	"License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
	"Natural Language :: English",
	"Operating System :: MacOS",
	"Operating System :: Microsoft :: Windows",
	"Operating System :: POSIX :: Linux",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Topic :: Utilities",
	"Topic :: Other/Nonlisted Topic",
	"Typing :: Typed"
]
description = "Helps sort and organize your comic collection by using the information stored in ComicInfo files."
include = [
	{ path = "tests", format = "sdist" }
]
keywords = ["comics", "comic", "metadata", "tagging", "tagger"]
license = "GPL-3.0-or-later"
maintainers = ["Buried-In-Code <BuriedInCode@tuta.io>"]
name = "Dex-Starr"
packages = [
	{ include = "dex_starr" }
]
readme = "README.md"
repository = "https://github.com/Buried-In-Code/Dex-Starr"
version = "0.2.0"

[tool.poetry.dependencies]
python = "^3.8"
jsonschema = "^4.4.0"
mokkari = "^2.1.0"
pathvalidate = "^2.5.0"
patool = "^1.12"
ratelimit = "^2.2.1"
requests = "^2.27.1"
rich = "^12.0.0"
"ruamel.yaml" = "^0.17.21"
simyan = "^0.7.4"
xmltodict = "^0.12.0"
yamale = "^4.0.3"
dataclasses-json = "^0.5.6"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
pytest = "^7.1.0"
pytest-cov = "^3.0.0"
tox = "^3.24.5"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.scripts]
Dex-Starr = 'dex_starr.__main__:main'

[tool.poetry.urls]
"Issue Tracker" = "https://github.com/Buried-In-Code/Dex-Starr/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[tool.isort]
default_section = "THIRDPARTY"
known_first_party = []
known_third_party = []
line_length = 100
multi_line_output = 3
profile = "black"

[tool.coverage.run]
source = ["dex_starr"]

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
addopts = ["--cov", "-x"]
